pyproject.toml
[project]
name = "licence-plate-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "google-cloud-storage>=3.4.0",
    "openai>=2.0.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ultralytics>=8.3.204",
    "uvicorn>=0.37.0",
]
